+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - [fix: fix lossless deregister failed when no healthcheck configured](https://github.com/Tencent/spring-cloud-tencent/pull/1345)
 - [feat:add zero protection.](https://github.com/Tencent/spring-cloud-tencent/pull/1346)
 - [fix:fix no registry when lossless is disabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1347)
-- [fix:fix the ratelimit bug for 2022](https://github.com/Tencent/spring-cloud-tencent/pull/1348)
+- [fix:fix the ratelimit bug for 2021](https://github.com/Tencent/spring-cloud-tencent/pull/1348)
 - [feat:add Tencent Cloud TSF support.](https://github.com/Tencent/spring-cloud-tencent/pull/1350)
 - [feat:support consul config.](https://github.com/Tencent/spring-cloud-tencent/pull/1352)
 - [feat:add trace report support.](https://github.com/Tencent/spring-cloud-tencent/pull/1353)
@@ -35,3 +35,4 @@
 - [feat:upgrade api circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/1440)
 - [feat: support lossless config from console & support warmup.](https://github.com/Tencent/spring-cloud-tencent/pull/1446)
 - [feat:add admin http handler.](https://github.com/Tencent/spring-cloud-tencent/pull/1449)
+- [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1456)